+++ fib_project/src/mock.cc
@@ -1,25 +1,11 @@
-#include<iostream>
-#include<string>
+#include "mock.h"
 
 
+// 1. Parameterized constructor of MyDataBase
+MyDataBase::MyDataBase(DataBaseConnect& _dbC): dbC(_dbC){}
 
-
-class DataBaseConnect{
-      public:
-      virtual bool login(std::string username, std::string password) { return true;}
-      virtual bool logout(std::string username) { return true;};
-      virtual int fetchRecord(){return -1;};
-};
-
-
-
-class MyDataBase{
-    DataBaseConnect &dbC;
-
-    public:
-    MyDataBase(DataBaseConnect &_dbC) : dbC(_dbC) {}
-
-    int Init(std::string username, std::string password){
+// 2. Init Method of class MyDataBase
+int MyDataBase::Init(std::string username, std::string password){
         if(dbC.login(username, password) != true){
             std::cout<<"DataBase Connection Failure !!!\n";
             return -1;
@@ -29,7 +15,15 @@ class MyDataBase{
         }
     }
 
-};
+
+
+
+
+
+
+
+
+
